1.增加Adapter模块文件 2.增加Pflash & PflashIf 3.修改Crc模块

+++ Bootloader/Rte/Rte_Queque.h
@@ -35,7 +35,7 @@ typedef struct
 
 typedef boolean Queque_Bool;
 typedef uint8	Queque_StdRet;
-typedef uint16	Queque_EleType;
+typedef uint8	Queque_EleType;
